+++ changelog.md
@@ -1,11 +1,15 @@
+# Version 1.2.1 (2022-10-13)
+
+– for-loop to reduce code-size
+
 # Version 1.2 (2022-10-13)
 
-– Auto-delete vorhandene Dateien
+– Auto-delete existing files
 
 # Version 1.1 (2022-09-30)
 
-– Sys-Exits hinzugefügt
-– Multi-Choise hinzugefügt
+– added sys-exits
+– added selection of multiple files
 – Failure-Prevention (aborts)
 
 # Version 1